general refactoring, still more to do
