+++ braille.txt
@@ -1,3 +1,1 @@
-..0.0.....0.
-..0.......0.
-.0.......0..
+Invalid Input, put something else
