+++ lib/dictionary.rb
@@ -1,52 +1,43 @@
+require_relative 'lookup_table'
 class Dictionary
+  include LookupTable
 
   attr_reader :text_input, :lookup_table, :input_braille
 
-  def initialize(input)
-    @text_input = split_text(input)
+  def initialize(text_input)
+    @text_input = split_text(text_input)
     @lookup_table = create_dictionary
-    @input_braille = is_braille?(input)
+    @input_braille = is_braille?(text_input)
+  end
+
+  def write_to_file
+    if @input_braille
+      read_characters
+    else
+      write_braille_to_file
+    end
   end
 
   def split_text(input)
-    if is_braille?(input)
-      braille_text = input.split('')
-      # require 'pry'; binding.pry
-      if braille_text[-1] == "\n"
-        return braille_text[0..-2]
-      end
-      return braille_text
-    else
-      text = input.split('')
-      if text[-1] == "\n"
-        return text[0..-2]
-      end
-      return text
+    text = input.split('')
+    if text[-1] == "\n"
+      return text[0..-2]
     end
+    return text
   end
 
   def text_to_rows
     return_count = 0
-    top_row = []
-    mid_row = []
-    bot_row = []
+    rows = [[],[],[]]
     @text_input.each do |letter|
       return_count += 1 if letter == "\n"
-      if return_count % 3 == 0
-        top_row << letter if letter != "\n"
-      elsif return_count % 3 == 1
-        mid_row << letter if letter != "\n"
-      else return_count % 3 == 2
-        bot_row << letter if letter != "\n"
-      end
+      rows[0] << letter if letter != "\n" if return_count % 3 == 0
+      rows[1] << letter if letter != "\n" if return_count % 3 == 1
+      rows[2] << letter if letter != "\n" if return_count % 3 == 2
     end
-    [top_row, mid_row, bot_row]
+    # require 'pry'; binding.pry
+    rows
   end
-
-  # def rows_to_character
-  #   braille_text = text_to_rows.join
-  #   @lookup_table.key(braille_text)
-  # end
 
   def read_characters
     message = ""
@@ -54,7 +45,6 @@ class Dictionary
     adjustable_text_to_rows = text_to_rows
     is_cap = false
     while adjustable_text_to_rows.flatten.size != 0 do
-      # require "pry"; binding.pry
       adjustable_text_to_rows.each do |row|
         letter << row[0]
         letter << row[1]
@@ -63,19 +53,16 @@ class Dictionary
       if @lookup_table.key(letter) == "cap"
         is_cap = true
       elsif is_cap == true
-        # require 'pry'; binding.pry
         message << @lookup_table.key(letter).upcase
         is_cap = false
       else
         message << @lookup_table.key(letter)
       end
-
       letter = ""
-
     end
-
     original_message = File.open("original_message.txt", "w")
     original_message.write(message)
+    original_message.close
     return message
   end
 
@@ -85,42 +72,6 @@ class Dictionary
 
   def create_r
     "0.\n00\n0."
-  end
-
-  def create_dictionary
-    hash = {}
-    hash["a"] = "0....."
-    hash["b"] = "0.0..."
-    hash["c"] = "00...."
-    hash["d"] = "00.0.."
-    hash["e"] = "0..0.."
-    hash["f"] = "000..."
-    hash["g"] = "0000.."
-
-    hash["h"] = "0.00.."
-    hash["i"] = ".00..."
-    hash["j"] = ".000.."
-    hash["k"] = "0...0."
-    hash["l"] = "0.0.0."
-    hash["m"] = "00..0."
-    hash["n"] = "00.00."
-    hash["o"] = "0..00."
-    hash["p"] = "000.0."
-
-    hash["q"] = "00000."
-    hash["r"] = "0.000."
-    hash["s"] = ".00.0."
-    hash["t"] = ".0000."
-    hash["u"] = "0...00"
-    hash["v"] = "0.0.00"
-    hash["w"] = ".000.0"
-    hash["x"] = "00..00"
-    hash["y"] = "00.000"
-    hash["z"] = "0..000"
-
-    hash[" "] = "......"
-    hash["cap"] = ".....0"
-    return hash
   end
 
   def write_brail(character)
@@ -135,40 +86,32 @@ class Dictionary
 
   def write_braille_to_file
     braille = File.open("braille.txt", "w")
-    top_row = ""
-    mid_row = ""
-    bot_row = ""
+    rows = ["","",""]
     count = 0
     @text_input.each do |character|
       if write_brail(character) == "Invalid Input"
         braille.write("Invalid Input, put something else")
-        break
+        return "Invalid Input"
       else
         count += 1
         if character == character.upcase && character != character.downcase
-          top_row += "#{write_brail("cap")[0]}#{write_brail("cap")[1]}"
-          mid_row += "#{write_brail("cap")[2]}#{write_brail("cap")[3]}"
-          bot_row += "#{write_brail("cap")[4]}#{write_brail("cap")[5]}"
+          rows[0] += "#{write_brail("cap")[0]}#{write_brail("cap")[1]}"
+          rows[1] += "#{write_brail("cap")[2]}#{write_brail("cap")[3]}"
+          rows[2] += "#{write_brail("cap")[4]}#{write_brail("cap")[5]}"
           count += 1
         end
-
-        top_row += "#{write_brail(character.downcase)[0]}#{write_brail(character.downcase)[1]}"
-        mid_row += "#{write_brail(character.downcase)[2]}#{write_brail(character.downcase)[3]}"
-        bot_row += "#{write_brail(character.downcase)[4]}#{write_brail(character.downcase)[5]}"
+        rows[0] += "#{write_brail(character.downcase)[0]}#{write_brail(character.downcase)[1]}"
+        rows[1] += "#{write_brail(character.downcase)[2]}#{write_brail(character.downcase)[3]}"
+        rows[2] += "#{write_brail(character.downcase)[4]}#{write_brail(character.downcase)[5]}"
         if count % 40 == 0
-          braille.write("#{top_row}\n")
-          braille.write("#{mid_row}\n")
-          braille.write("#{bot_row}\n")
-          top_row = ""
-          mid_row = ""
-          bot_row = ""
+          braille.write("#{rows[0]}\n#{rows[1]}\n#{rows[2]}\n")
+          rows = ["","",""]
         end
       end
     end
-      braille.write("#{top_row}\n")
-      braille.write("#{mid_row}\n")
-      braille.write("#{bot_row}")
+    braille.write("#{rows[0]}\n#{rows[1]}\n#{rows[2]}")
     braille.close
+    return rows
   end
 
 end
