+++ original_message.txt
@@ -1,1 +1,1 @@
- rllo this is a longer message for practice purposes
+Ba B
